Foreign key adjustment referring to order_product

+++ database/migrations/2023_05_02_224517_create_orders_products_table.php
@@ -13,14 +13,14 @@ return new class extends Migration
     {
         Schema::create('order_product', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('product_id')->unsigned()->nullable();
+            $table->bigInteger('product_id')->unsigned()->nullable();
             $table->foreign('product_id')
-              ->references('id')
-              ->on('products')->onDelete('cascade');
-            $table->integer('order_id')->unsigned()->nullable();
+                ->references('id')
+                ->on('products')->onDelete('cascade');
+            $table->bigInteger('order_id')->unsigned()->nullable();
             $table->foreign('order_id')
-              ->references('id')
-              ->on('orders')->onDelete('cascade');
+                ->references('id')
+                ->on('orders')->onDelete('cascade');
             $table->integer('qtd')->nullable();
             $table->decimal('price', $precision = 10, $scale = 2)->nullable();
             $table->timestamps();
